refactor: draftAttributeRenderer
