+++ packages/renderers/lib/src/renderers/request_item_renderer/widgets/processed_query_renderer.dart
@@ -34,7 +34,6 @@ class ProcessedIdentityAttributeQueryRenderer extends StatelessWidget {
     }
     return IdentityAttributeValueRenderer(
       value: query.results.first.value as IdentityAttributeValue,
-      controller: controller,
       onEdit: onEdit,
     );
   }
